Calculator 0.5가 나오게 수정

+++ src/java1005/Calculator.java
@@ -14,6 +14,6 @@ public class Calculator {
     }
 
     public void divide(int a, int b){
-        System.out.println(a / (double) b);
+        System.out.println(a / (float) b);
     }
 }
